change extension name to withClickablePart from withSpacialPart.
