+++ lib/super_extensions.dart
@@ -1,4 +1,5 @@
 export 'src/context_extensions.dart';
 export 'src/number_extensions.dart';
 export 'src/string_extensions.dart';
-export 'src/text_widget_extensions.dart';
+export 'src/text_widget_extensions/models/part.dart';
+export 'src/text_widget_extensions/text_widget_extensions.dart';
